Renamed and cleaned up the package.
