+++ README
@@ -1,3 +1,3 @@
-ROS Driver for XSens MT/MTi/MTi-G devices.
+ROS 2 driver for XSens MT/MTi/MTi-G devices.
 
 See: http://ros.org/wiki/ethzasl_xsens_driver
