+++ mainpage.dox
@@ -2,11 +2,11 @@
 \mainpage
 \htmlinclude manifest.html
 
-The \b xsens_driver package provides \b mtnode.py, a generic ROS node publishing the data streamed by an XSens motion tracker (MT, MTi, MTi-G...) from third, fourth or fifth generation.
+The \b xsens_driver package provides \b mtnode.py, a generic ROS 2 node publishing the data streamed by an XSens motion tracker (MT, MTi, MTi-G...) from third, fourth or fifth generation.
 
 \section topics Topics
 
-The ROS node is a wrapper around the \b mtdevice::MTDevice class. It can publish the following topics, depending on the configuration of the device:
+The ROS 2 node is a wrapper around the \b mtdevice::MTDevice class. It can publish the following topics, depending on the configuration of the device:
 - \c /imu/data (<a href="http://www.ros.org/doc/api/sensor_msgs/html/msg/Imu.html">sensor_msgs::Imu</a>): orientation, angular velocity, linear acceleration,
 - \c /fix (<a href="http://www.ros.org/doc/api/sensor_msgs/html/msg/NavSatFix.html">sensor_msgs::NavSatFix</a>): longitude, latitude, altitude (from filtered/fused GPS and IMU),
 - \c /raw_fix (<a href="http://www.ros.org/doc/api/sensor_msgs/html/msg/NavSatFix.html">sensor_msgs::NavSatFix</a>): longitude, latitude, altitude (from ONLY GPS),
@@ -29,16 +29,16 @@ The covariance information in the <a href="http://www.ros.org/doc/api/sensor_msg
 \section params Parameters
 
 The nodes can take the following parameters:
-- \a ~device (\c auto): the path of the device file to connect to the imu; \c auto will look through all serial devices to find the first one.
-- \a ~baudrate (0): the baudrate of the IMU (unused for \c auto device); 0 will try to auto-detect baudrate.
-- \a ~timeout (0.002): the timeout for communication with the IMU.
-- \a ~initial_wait (0.1): initial wait to allow the device to come up before trying to communicate.
-- \a ~frame_id (\c /base_imu): the frame id of the IMU.
-- \a ~frame_local (\c ENU): the desired frame orientation (\c ENU, \c NED, or \c NWU).
-- \a ~no_rotation_duration (0): the int duration in seconds of the no-rotation calibration procedure at the start of the node (only for mark iv devices). If 0, the procedure is not launched.
-- \a ~angular_velocity_covariance_diagonal ([0.0004, 0.0004, 0.0004]) diagonal elements of the covariance matrix on the angular velocity published on the \c /imu/data topic.
-- \a ~linear_acceleration_covariance_diagonal ([0.0004, 0.0004, 0.0004]) diagonal elements of the covariance matrix on the linear acceleration published on the \c /imu/data topic.
-- \a ~orientation_covariance_diagonal ([0.01745, 0.01745, 0.15708]) diagonal elements of the covariance matrix on the orientation published on the \c /imu/data topic.
+- \a device (\c auto): the path of the device file to connect to the imu; \c auto will look through all serial devices to find the first one.
+- \a baudrate (0): the baudrate of the IMU (unused for \c auto device); 0 will try to auto-detect baudrate.
+- \a timeout (0.002): the timeout for communication with the IMU.
+- \a initial_wait (0.1): initial wait to allow the device to come up before trying to communicate.
+- \a frame_id (\c /base_imu): the frame id of the IMU.
+- \a frame_local (\c ENU): the desired frame orientation (\c ENU, \c NED, or \c NWU).
+- \a no_rotation_duration (0): the int duration in seconds of the no-rotation calibration procedure at the start of the node (only for mark iv devices). If 0, the procedure is not launched.
+- \a angular_velocity_covariance_diagonal ([0.0004, 0.0004, 0.0004]) diagonal elements of the covariance matrix on the angular velocity published on the \c /imu/data topic.
+- \a linear_acceleration_covariance_diagonal ([0.0004, 0.0004, 0.0004]) diagonal elements of the covariance matrix on the linear acceleration published on the \c /imu/data topic.
+- \a orientation_covariance_diagonal ([0.01745, 0.01745, 0.15708]) diagonal elements of the covariance matrix on the orientation published on the \c /imu/data topic.
 
 \section dialout Permissions
