+++ package.xml
@@ -1,10 +1,10 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>ethzasl_xsens_driver_ros2</name>
-  <version>0.0.1</version>
-  <description>ROS 2 Driver for XSens MT/MTi/MTi-G devices.</description>
-  <maintainer email="dominic.baril@norlab.ulaval.ca">Dominic Baril</maintainer>
+  <name>xsens_driver</name>
+  <version>3.0.0</version>
+  <description>ROS 2 driver for XSens MT/MTi/MTi-G devices.</description>
+  <maintainer email="francis.colas@mavt.ethz.ch">Francis Colas</maintainer>
   <license>BSD</license>
 
   <depend>rclpy</depend>
