+++ setup.cfg
@@ -1,5 +1,5 @@
 [develop]
-script_dir=$base/lib/ethzasl_xsens_driver_ros2
+script_dir=$base/lib/xsens_driver
 [install]
-install_scripts=$base/lib/ethzasl_xsens_driver_ros2
+install_scripts=$base/lib/xsens_driver
